refactor: Remove code duplication
