+++ packages/react-native-executorch/common/rnexecutorch/models/embeddings/BaseEmbeddings.h
@@ -1,0 +1,17 @@
+#pragma once
+
+#include <rnexecutorch/models/BaseModel.h>
+
+namespace rnexecutorch {
+
+class BaseEmbeddings : public BaseModel {
+public:
+  BaseEmbeddings(const std::string &modelSource,
+                 std::shared_ptr<react::CallInvoker> callInvoker);
+
+protected:
+  std::shared_ptr<OwningArrayBuffer>
+  postprocess(const Result<std::vector<EValue>> &forwardResult);
+};
+
+}; // namespace rnexecutorch
