+++ packages/react-native-executorch/common/rnexecutorch/models/embeddings/image/ImageEmbeddings.h
@@ -4,22 +4,19 @@
 #include <executorch/runtime/core/evalue.h>
 #include <opencv2/opencv.hpp>
 
-#include <rnexecutorch/models/BaseModel.h>
-#include <span>
+#include <rnexecutorch/models/embeddings/BaseEmbeddings.h>
 
 namespace rnexecutorch {
 using executorch::extension::TensorPtr;
 using executorch::runtime::EValue;
 
-class ImageEmbeddings : public BaseModel {
+class ImageEmbeddings : public BaseEmbeddings {
 public:
   ImageEmbeddings(const std::string &modelSource,
                   std::shared_ptr<react::CallInvoker> callInvoker);
-  std::shared_ptr<OwningArrayBuffer> generate(std::string imageSource);
+  std::shared_ptr<OwningArrayBuffer> generate(const std::string imageSource);
 
 private:
-  std::shared_ptr<OwningArrayBuffer> postprocess(std::span<float> modelOutput);
-
   cv::Size modelImageSize{0, 0};
 };
