+++ packages/react-native-executorch/common/rnexecutorch/models/embeddings/text/TextEmbeddings.cpp
@@ -9,7 +9,7 @@ using namespace executorch::extension;
 TextEmbeddings::TextEmbeddings(const std::string &modelSource,
                                const std::string &tokenizerSource,
                                std::shared_ptr<react::CallInvoker> callInvoker)
-    : BaseModel(modelSource, callInvoker),
+    : BaseEmbeddings(modelSource, callInvoker),
       tokenizer(
           std::make_unique<TokenizerModule>(tokenizerSource, callInvoker)) {}
 
@@ -36,41 +36,26 @@ TokenIdsWithAttentionMask TextEmbeddings::preprocess(const std::string &input) {
 std::shared_ptr<OwningArrayBuffer>
 TextEmbeddings::generate(const std::string input) {
   auto preprocessed = preprocess(input);
+
   std::vector<int32_t> tokenIdsShape = {
       1, static_cast<int32_t>(preprocessed.inputIds.size())};
   std::vector<int32_t> attnMaskShape = {
       1, static_cast<int32_t>(preprocessed.attentionMask.size())};
+
   auto tokenIds = make_tensor_ptr(tokenIdsShape, preprocessed.inputIds.data(),
                                   ScalarType::Long);
   auto attnMask = make_tensor_ptr(
       attnMaskShape, preprocessed.attentionMask.data(), ScalarType::Long);
+
   auto forwardResult = BaseModel::forward({tokenIds, attnMask});
+
   if (!forwardResult.ok()) {
     throw std::runtime_error(
-        "Failed to forward, error: " +
+        "Function forward in TextEmbeddings failed with error code: " +
         std::to_string(static_cast<uint32_t>(forwardResult.error())));
   }
 
-  auto forwardResultTensor = forwardResult->at(0).toTensor();
-  auto dataPtr = forwardResultTensor.mutable_data_ptr();
-  auto outputNumel = forwardResultTensor.numel();
-
-  std::span<float> modelOutputSpan(static_cast<float *>(dataPtr), outputNumel);
-  std::span<const int64_t> attnMaskSpan(preprocessed.attentionMask.data(),
-                                        preprocessed.attentionMask.size());
-
-  return postprocess(modelOutputSpan);
-}
-
-std::shared_ptr<OwningArrayBuffer>
-TextEmbeddings::postprocess(std::span<float> modelOutput) {
-  auto createBuffer = [](const auto &data, size_t size) {
-    auto buffer = std::make_shared<OwningArrayBuffer>(size);
-    std::memcpy(buffer->data(), data, size);
-    return buffer;
-  };
-
-  return createBuffer(modelOutput.data(), modelOutput.size_bytes());
+  return BaseEmbeddings::postprocess(forwardResult);
 }
 
 } // namespace rnexecutorch
