+++ packages/react-native-executorch/common/rnexecutorch/models/embeddings/text/TextEmbeddings.h
@@ -1,8 +1,7 @@
 #pragma once
 
 #include <rnexecutorch/TokenizerModule.h>
-#include <rnexecutorch/models/BaseModel.h>
-#include <span>
+#include <rnexecutorch/models/embeddings/BaseEmbeddings.h>
 
 namespace rnexecutorch {
 
@@ -11,7 +10,7 @@ struct TokenIdsWithAttentionMask {
   std::vector<int64_t> attentionMask;
 };
 
-class TextEmbeddings : public BaseModel {
+class TextEmbeddings : public BaseEmbeddings {
 public:
   TextEmbeddings(const std::string &modelSource,
                  const std::string &tokenizerSource,
@@ -21,7 +20,7 @@ public:
 private:
   std::vector<std::vector<int32_t>> inputShapes;
   TokenIdsWithAttentionMask preprocess(const std::string &input);
-  std::shared_ptr<OwningArrayBuffer> postprocess(std::span<float> modelOutput);
   std::unique_ptr<TokenizerModule> tokenizer;
 };
+
 }; // namespace rnexecutorch
